alphavantage_scraper/pyproject.toml
[tool.poetry]
name = "av-scraper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "av_scraper"}]

[tool.poetry.dependencies]
python = "^3.10"
influxdb-client = "^1.34.0"
python-dotenv = "^0.21.0"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
